refactor (EndPointUtil) split off ErrorUtil
