+++ lib/Endpoint.js
@@ -1,7 +1,7 @@
 const ApiRequest = require('./ApiRequest');
 
 const ParameterError = require('./errors/ParameterError');
-const EndpointUtil = require('./util/EndpointUtil');
+const ErrorUtil = require('./util/ErrorUtil');
 const RegionAndPlatformUtil = require('./util/RegionAndPlatformUtil');
 
 /**
@@ -14,7 +14,7 @@ class Endpoint {
 	 */
 	constructor(name, config, rateLimiter) {
 		if (rateLimiter){
-			EndpointUtil.throwIfNotRateLimiter(rateLimiter, 'rateLimiter');
+			ErrorUtil.throwIfNotRateLimiter(rateLimiter, 'rateLimiter');
 		}
 		if (arguments.length < 2) {
 			throw new ParameterError('Endpoint expects 2 mandatory Parameter and 1 optional Parameter: ' +
