+++ lib/endpoints/ChampionEndpoint.js
@@ -1,6 +1,7 @@
 const Endpoint = require('../Endpoint');
 const EndpointUtil = require('../util/EndpointUtil');
 const ParameterUtil = require('../util/ParameterUtil');
+const ErrorUtil = require('../util/ErrorUtil');
 
 /***
  * Endpoint to receive client information about champions on the league platform
@@ -22,8 +23,8 @@ class ChampionEndpoint extends Endpoint {
 	gettingList(platformIdOrRegion, {freeToPlay = false} = {}) {
 		const {_platformId, _options} =
 			ParameterUtil.extractPlatformIdAndOptions(platformIdOrRegion,{freeToPlay});
-		EndpointUtil.throwIfNotBoolean(freeToPlay, 'freeToPlay');
-		return this.executingRequest(``, _platformId, EndpointUtil.buildQueryStringFromOptions(_options));
+		ErrorUtil.throwIfNotBoolean(freeToPlay, 'freeToPlay');
+		return this.executingRequest(``, _platformId, EndpointUtil.buildQueryStringFromOptions(_options)); // TODO: move into Endpoint class (executing...)
 	}
 
 	/**
@@ -33,7 +34,7 @@ class ChampionEndpoint extends Endpoint {
 	 * @return {Bluebird<ChampionDto>}
 	 */
 	gettingById(championId, platformIdOrRegion = this.config.PLATFORM_ID) {
-		EndpointUtil.throwIfNotNumerical(championId, 'championId');
+		ErrorUtil.throwIfNotNumerical(championId, 'championId');
 		return this.executingRequest(`/${championId}`, platformIdOrRegion);
 	}
 }
