+++ lib/endpoints/ChampionMasteryEndpoint.js
@@ -1,5 +1,5 @@
 const Endpoint = require ('../Endpoint');
-const EndpointUtil = require('../util/EndpointUtil');
+const ErrorUtil = require('../util/ErrorUtil');
 
 /**
  * Endpoint to receive ChampionMastery information
@@ -18,7 +18,7 @@ class ChampionMasteryEndpoint extends Endpoint {
 	 * @return {Bluebird<ChampionMasteryDTO[]>}
 	 */
 	gettingBySummoner(summonerId, platformIdOrRegion = this.config.PLATFORM_ID){
-		EndpointUtil.throwIfNotNumerical(summonerId, 'summonerId');
+		ErrorUtil.throwIfNotNumerical(summonerId, 'summonerId');
 		return this.executingRequest(`/champion-masteries/by-summoner/${summonerId}`,  platformIdOrRegion);
 	}
 
@@ -30,8 +30,8 @@ class ChampionMasteryEndpoint extends Endpoint {
 	 * @return {Bluebird<ChampionMasteryDTO>}
 	 */
 	gettingBySummonerForChampion(summonerId, championId, platformIdOrRegion = this.config.PLATFORM_ID){
-		EndpointUtil.throwIfNotNumerical(summonerId, 'summonerId');
-		EndpointUtil.throwIfNotNumerical(championId, 'championId');
+		ErrorUtil.throwIfNotNumerical(summonerId, 'summonerId');
+		ErrorUtil.throwIfNotNumerical(championId, 'championId');
 		return this.executingRequest(`/champion-masteries/by-summoner/${summonerId}/by-champion/${championId}`,  platformIdOrRegion);
 	}
 
@@ -39,7 +39,7 @@ class ChampionMasteryEndpoint extends Endpoint {
 	 * Returns the ChampionMastery score (combined champion-levels for the summoner)
 	 * @return {Bluebird<number>}*/
 	gettingScoresBySummoner(summonerId, platformIdOrRegion = this.config.PLATFORM_ID){
-		EndpointUtil.throwIfNotNumerical(summonerId, 'summonerId');
+		ErrorUtil.throwIfNotNumerical(summonerId, 'summonerId');
 		return this.executingRequest(`/scores/by-summoner/${summonerId}`,  platformIdOrRegion);
 	}
 }
