+++ lib/endpoints/LeagueEndpoint.js
@@ -1,5 +1,5 @@
 const Endpoint = require ('../Endpoint');
-const EndpointUtil = require('../util/EndpointUtil');
+const ErrorUtil = require('../util/ErrorUtil');
 
 /***
  * Endpoint to receive ranked information about leagues by queue or for specific summoners
@@ -18,7 +18,7 @@ class LeagueEndpoint extends Endpoint {
 	 * @return {Bluebird<LeagueListDTO>}
 	 */
 	gettingChallengerLeague( rankedQueueConfigId, platformIdOrRegion = this.config.PLATFORM_ID){
-		EndpointUtil.throwIfRankedQueueConfigIdInvalid(rankedQueueConfigId);
+		ErrorUtil.throwIfRankedQueueConfigIdInvalid(rankedQueueConfigId);
 		return this.executingRequest(`/challengerleagues/by-queue/${rankedQueueConfigId}`, platformIdOrRegion);
 	}
 
@@ -29,7 +29,7 @@ class LeagueEndpoint extends Endpoint {
 	 * @return {Bluebird<LeagueListDTO>}
 	 */
 	gettingMasterLeague( rankedQueueConfigId, platformIdOrRegion = this.config.PLATFORM_ID){
-		EndpointUtil.throwIfRankedQueueConfigIdInvalid(rankedQueueConfigId);
+		ErrorUtil.throwIfRankedQueueConfigIdInvalid(rankedQueueConfigId);
 		return this.executingRequest(`/masterleagues/by-queue/${rankedQueueConfigId}`,  platformIdOrRegion);
 	}
 
@@ -38,7 +38,7 @@ class LeagueEndpoint extends Endpoint {
 	 * @return {Bluebird<LeagueListDTO[]>}
 	 */
 	gettingLeagueForSummonerId( summonerId, platformIdOrRegion = this.config.PLATFORM_ID){
-		EndpointUtil.throwIfNotNumerical(summonerId, 'summonerId');
+		ErrorUtil.throwIfNotNumerical(summonerId, 'summonerId');
 		return this.executingRequest(`/leagues/by-summoner/${summonerId}`,  platformIdOrRegion);
 	}
 
@@ -47,7 +47,7 @@ class LeagueEndpoint extends Endpoint {
 	 * @return {Bluebird<LeaguePositionDTO[]>}
 	 */
 	gettingPositionsForSummonerId( summonerId, platformIdOrRegion = this.config.PLATFORM_ID){
-		EndpointUtil.throwIfNotNumerical(summonerId, 'summonerId');
+		ErrorUtil.throwIfNotNumerical(summonerId, 'summonerId');
 		return this.executingRequest(`/positions/by-summoner/${summonerId}`,  platformIdOrRegion);
 	}
