+++ lib/endpoints/MasteriesEndpoint.js
@@ -1,5 +1,5 @@
 const Endpoint = require ('../Endpoint');
-const EndpointUtil = require('../util/EndpointUtil');
+const ErrorUtil = require('../util/ErrorUtil');
 
 /***
  * Endpoint to receive Masteries of a player
@@ -18,7 +18,7 @@ class MasteriesEndpoint extends Endpoint {
 	 * @return {Bluebird<MasteryPagesDto>}
 	 */
 	gettingBySummoner( summonerId, platformIdOrRegion = this.config.PLATFORM_ID){
-		EndpointUtil.throwIfNotNumerical(summonerId, 'summonerId');
+		ErrorUtil.throwIfNotNumerical(summonerId, 'summonerId');
 		return this.executingRequest(`/masteries/by-summoner/${summonerId}`,  platformIdOrRegion);
 	}
 }
