+++ lib/endpoints/MatchEndpoint.js
@@ -1,6 +1,7 @@
 const Endpoint = require('../Endpoint');
 const EndpointUtil = require('../util/EndpointUtil');
 const ParameterUtil = require('../util/ParameterUtil');
+const ErrorUtil = require('../util/ErrorUtil');
 
 /***
  * Endpoint to receive information about finished Match
@@ -19,7 +20,7 @@ class MatchEndpoint extends Endpoint {
 	 * @return {Bluebird<MatchDto>}
 	 */
 	gettingById(gameId, platformIdOrRegion = this.config.PLATFORM_ID) {
-		EndpointUtil.throwIfNotNumerical(gameId, 'gameId');
+		ErrorUtil.throwIfNotNumerical(gameId, 'gameId');
 		return this.executingRequest(`/matches/${gameId}`, platformIdOrRegion);
 	}
 
@@ -30,7 +31,7 @@ class MatchEndpoint extends Endpoint {
 	 * @return {Bluebird<MatchTimelineDto>}
 	 */
 	gettingTimelineById(gameId, platformIdOrRegion = this.config.PLATFORM_ID) {
-		EndpointUtil.throwIfNotNumerical(gameId, 'gameId');
+		ErrorUtil.throwIfNotNumerical(gameId, 'gameId');
 		return this.executingRequest(`/timelines/by-match/${gameId}`, platformIdOrRegion);
 	}
 
@@ -52,7 +53,7 @@ class MatchEndpoint extends Endpoint {
 	 * @return {Bluebird<MatchlistDto>}
 	 */
 	gettingListByAccount(accountId, platformIdOrRegion, options = {}) {
-		EndpointUtil.throwIfNotNumerical(accountId, 'accountId');
+		ErrorUtil.throwIfNotNumerical(accountId, 'accountId');
 		const {_platformId, _options} = ParameterUtil.extractPlatformIdAndOptions(platformIdOrRegion, options);
 		return this.executingRequest(`/matchlists/by-account/${accountId}`, _platformId, EndpointUtil.buildQueryStringFromOptions(_options));
 	}
@@ -64,7 +65,7 @@ class MatchEndpoint extends Endpoint {
 	 * @return {Bluebird<MatchlistDto>}
 	 */
 	gettingRecentListByAccount(accountId, platformIdOrRegion = this.config.PLATFORM_ID) {
-		EndpointUtil.throwIfNotNumerical(accountId, 'accountId');
+		ErrorUtil.throwIfNotNumerical(accountId, 'accountId');
 		return this.executingRequest(`/matchlists/by-account/${accountId}/recent`, platformIdOrRegion);
 	}
 
@@ -90,7 +91,7 @@ class MatchEndpoint extends Endpoint {
 	 * @return {Bluebird<MatchDto>}
 	 */
 	gettingByIdForTournament(gameId, tournamentCode, platformIdOrRegion = this.config.PLATFORM_ID) { // TODO: unit tests!?
-		EndpointUtil.throwIfNotNumerical(gameId, 'gameId');
+		ErrorUtil.throwIfNotNumerical(gameId, 'gameId');
 		if(typeof tournamentCode !== 'string'){
 			throw new TypeError(`tournamentCode should be a string. Received: ${tournamentCode}`);
 		}
