+++ lib/endpoints/RunesEndpoint.js
@@ -1,5 +1,5 @@
 const Endpoint = require ('../Endpoint');
-const EndpointUtil = require('../util/EndpointUtil');
+const ErrorUtil = require('../util/ErrorUtil');
 
 /***
  * Endpoint to receive Masteries of a player
@@ -18,7 +18,7 @@ class RunesEndpoint extends Endpoint {
 	 * @return {Bluebird<RunePagesDto>}
 	 */
 	gettingBySummoner( summonerId, platformIdOrRegion = this.config.PLATFORM_ID){
-		EndpointUtil.throwIfNotNumerical(summonerId, 'summonerId');
+		ErrorUtil.throwIfNotNumerical(summonerId, 'summonerId');
 		return this.executingRequest(`/runes/by-summoner/${summonerId}`,  platformIdOrRegion);
 	}
 }
