+++ lib/endpoints/SpectatorEndpoint.js
@@ -1,5 +1,5 @@
 const Endpoint = require ('../Endpoint');
-const EndpointUtil = require('../util/EndpointUtil');
+const ErrorUtil = require('../util/ErrorUtil');
 
 /***
  * Endpoint to receive live match information
@@ -18,7 +18,7 @@ class SpectatorEndpoint extends Endpoint {
 	 * @return {Bluebird<CurrentGameInfo>}
 	 */
 	gettingActiveGame( summonerId, platformIdOrRegion = this.config.PLATFORM_ID){
-		EndpointUtil.throwIfNotNumerical(summonerId, 'summonerId');
+		ErrorUtil.throwIfNotNumerical(summonerId, 'summonerId');
 		return this.executingRequest(`/active-games/by-summoner/${summonerId}`,  platformIdOrRegion);
 	}
