+++ lib/endpoints/StaticDataEndpoint.js
@@ -1,6 +1,7 @@
 const Endpoint = require('../Endpoint');
 const EndpointUtil = require('../util/EndpointUtil');
 const ParameterUtil = require('../util/ParameterUtil');
+const ErrorUtil = require('../util/ErrorUtil');
 
 // const ParameterError = require('../errors/ParameterError');
 
@@ -123,7 +124,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * @return {Bluebird<ChampionDto>}
 	 */
 	gettingChampionById(championId, platformIdOrRegion, options = {}) {
-		EndpointUtil.throwIfNotNumerical(championId, 'championId');
+		ErrorUtil.throwIfNotNumerical(championId, 'championId');
 		const {_platformId, _options} = ParameterUtil.extractPlatformIdAndOptions(platformIdOrRegion, options);
 		return this.executingRequest(`/champions/${championId}`, _platformId, EndpointUtil.buildQueryStringFromOptions(_options));
 	}
@@ -243,7 +244,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * @return {Bluebird<ItemDto>}
 	 */
 	gettingItemById(itemId, platformIdOrRegion, options = {}) {
-		EndpointUtil.throwIfNotNumerical(itemId, 'itemId');
+		ErrorUtil.throwIfNotNumerical(itemId, 'itemId');
 		const {_platformId, _options} = ParameterUtil.extractPlatformIdAndOptions(platformIdOrRegion, options);
 		return this.executingRequest(`/items/${itemId}`, _platformId, EndpointUtil.buildQueryStringFromOptions(_options));
 	}
@@ -523,7 +524,7 @@ class StaticDataEndpoint extends Endpoint {
 	 * @return {Bluebird<RuneDto>}
 	 */
 	gettingRunesById(runeId, platformIdOrRegion, options = {}) {
-		EndpointUtil.throwIfNotNumerical(runeId);
+		ErrorUtil.throwIfNotNumerical(runeId);
 		const {_platformId, _options} = ParameterUtil.extractPlatformIdAndOptions(platformIdOrRegion, options);
 		return this.executingRequest(`/runes/${runeId}`, _platformId, EndpointUtil.buildQueryStringFromOptions(_options));
 	}
