+++ lib/endpoints/SummonerEndpoint.js
@@ -1,5 +1,5 @@
 const Endpoint = require ('../Endpoint');
-const EndpointUtil = require('../util/EndpointUtil');
+const ErrorUtil = require('../util/ErrorUtil');
 const SummonerUtil = require('../util/SummonerUtil');
 
 /**
@@ -20,8 +20,8 @@ class SummonerEndpoint extends Endpoint {
 	 * @return {Bluebird<SummonerDto>}
      */
     gettingByName(name, platformIdOrRegion = this.config.PLATFORM_ID) {
-		EndpointUtil.throwIfParamIsArray(name, 'name');
-		EndpointUtil.throwIfNameIsInvalid(name, 'name');
+		ErrorUtil.throwIfIsArray(name, 'name');
+		ErrorUtil.throwIfNameIsInvalid(name, 'name');
 
 		let normalizedName = SummonerUtil.normalizeSummonerName(name);
         return this.executingRequest(`/by-name/${normalizedName}`,  platformIdOrRegion);
@@ -34,7 +34,7 @@ class SummonerEndpoint extends Endpoint {
 	 * @return {Bluebird<SummonerDto>}
 	 */
     gettingByAccount(accountId, platformIdOrRegion = this.config.PLATFORM_ID) {
-		EndpointUtil.throwIfNotNumerical(accountId, 'accountId');
+		ErrorUtil.throwIfNotNumerical(accountId, 'accountId');
 		return this.executingRequest(`/by-account/${accountId}`,  platformIdOrRegion);
 	}
 
@@ -45,7 +45,7 @@ class SummonerEndpoint extends Endpoint {
 	 * @return {Bluebird<SummonerDto>}
 	 */
     gettingById(summonerId, platformIdOrRegion = this.config.PLATFORM_ID) {
-		EndpointUtil.throwIfNotNumerical(summonerId, 'summonerId');
+		ErrorUtil.throwIfNotNumerical(summonerId, 'summonerId');
 		return this.executingRequest(`/${summonerId}`,  platformIdOrRegion);
 	}
 }
